some improvements and new versions in build.gradle.kts

+++ app/src/main/java/com/example/expensetracker/presentation/home/SwipeConfirm.kt
@@ -135,7 +135,7 @@ private fun DraggableControl(
                         note = note.value,
                         date = convertLocalDateToDate(date.value),
                         value = expense.value!!,
-                        categoryId = category.value!!.categoryId.toInt()
+                        categoryId = category.value!!.categoryId
                     )
                 )
                 bottomSheetViewModel.setBottomSheetExpanded(false)
